Moved autocomplete components into web/react/components/suggestion
