+++ web/react/components/search_bar.jsx
@@ -5,10 +5,10 @@ import * as client from '../utils/client.jsx';
 import * as AsyncClient from '../utils/async_client.jsx';
 import SearchStore from '../stores/search_store.jsx';
 import AppDispatcher from '../dispatcher/app_dispatcher.jsx';
-import SuggestionBox from '../components/suggestion_box.jsx';
-import SearchChannelProvider from '../components/search_channel_provider.jsx';
-import SearchSuggestionList from '../components/search_suggestion_list.jsx';
-import SearchUserProvider from '../components/search_user_provider.jsx';
+import SuggestionBox from './suggestion/suggestion_box.jsx';
+import SearchChannelProvider from './suggestion/search_channel_provider.jsx';
+import SearchSuggestionList from './suggestion/search_suggestion_list.jsx';
+import SearchUserProvider from './suggestion/search_user_provider.jsx';
 import * as utils from '../utils/utils.jsx';
 import Constants from '../utils/constants.jsx';
 var ActionTypes = Constants.ActionTypes;
