+++ web/react/components/suggestion/search_suggestion_list.jsx
@@ -1,9 +1,9 @@
 // Copyright (c) 2015 Mattermost, Inc. All Rights Reserved.
 // See License.txt for license information.
 
-import Constants from '../utils/constants.jsx';
+import Constants from '../../utils/constants.jsx';
 import SuggestionList from './suggestion_list.jsx';
-import * as Utils from '../utils/utils.jsx';
+import * as Utils from '../../utils/utils.jsx';
 
 export default class SearchSuggestionList extends SuggestionList {
     componentDidUpdate(prevProps, prevState) {
